+++ packages/cv/lib/utils/value_utils.dart
@@ -1,0 +1,1 @@
+export 'package:cv/src/utils.dart' show basicTypeCast, isBasicTypeOrNull;